requirements.txt
streamlit
youtube-transcript-api
yt-dlp
openai-whisper
ollama
fpdf2
pandas
spacy
fuzzywuzzy
python-Levenshtein
requests